requirements.txt
requests
django
pytest
pytest-django
django-filter
djangorestframework
djangorestframework-simplejwt
